requirements.txt
# Core Flask and Extensions
Flask==2.2.5
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.3.0
Flask-RESTful==0.3.10
Flask-DebugToolbar==0.13.1

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.11
alembic==1.12.0
greenlet==3.2.4

# Authentication & Security
bcrypt==4.0.1
PyJWT>=2.9.0,<3.0.0
WTForms==3.0.1
email-validator==2.0.0
python-dotenv==1.0.0

# Templates & Markup
Jinja2==3.1.2
MarkupSafe==2.1.3
Mako==1.3.10

# HTTP & Networking
Werkzeug==2.2.3
requests==2.32.3
urllib3==2.0.7
idna==3.11
certifi==2024.12.14
charset-normalizer==3.4.1

# Rate Limiting & Caching
limits==5.6.0
ordered-set==4.1.0
redis==5.3.0

# Development & Debugging
click==8.3.0
itsdangerous==2.2.0
blinker==1.9.0
rich==13.9.4
Pygments==2.19.2
python-magic==0.4.27
pillow==12.2.0

# Testing
pytest==8.3.5
pytest-cov==6.0.0
pytest-flask==1.3.0
factory-boy==3.3.3
Faker==38.0.0

# Production Server
gunicorn==25.3.0

# Utilities
aniso8601==10.0.1
Deprecated==1.2.18
dnspython==2.8.0
packaging==25.0
pytz==2025.2
six==1.17.0
typing_extensions==4.15.0
wrapt==1.17.3
markdown-it-py==4.0.0
mdurl==0.1.2